+++ src/components/ImgWrapper/index.js
@@ -1,0 +1,2 @@
+import ImgWrapper from './ImgWrapper';
+export default ImgWrapper;
