+++ src/components/Layout.jsx
@@ -1,5 +1,5 @@
 import { Outlet } from 'react-router-dom';
-import Header from 'components/Header/Header';
+import Header from 'components/Header';
 
 const Layout = () => {
   return (
