+++ src/components/MenuItem/index.js
@@ -1,0 +1,2 @@
+import MenuItem from './MenuItem';
+export default MenuItem;
