+++ src/components/ProductContainer/index.js
@@ -1,0 +1,2 @@
+import ProductContainer from './ProductContainer';
+export default ProductContainer;
