+++ src/components/ProductDetail/index.js
@@ -1,0 +1,2 @@
+import ProductDetail from './ProductDetail';
+export default ProductDetail;
