+++ src/components/ProductItem/ProductItem.jsx
@@ -38,23 +38,19 @@ const Styled = {
     width: 200px;
     cursor: pointer;
   `,
-
   ProductImage: styled.img`
     max-width: 100%;
   `,
-
   ProductDetail: styled.div`
     display: flex;
     justify-content: space-between;
     padding: 5px;
     width: 100%;
   `,
-
   ProductInfo: styled.div`
     display: flex;
     flex-direction: column;
   `,
-
   ProductName: styled.span`
     font-size: 14px;
     width: 147px;
@@ -63,17 +59,14 @@ const Styled = {
     overflow: hidden;
     text-overflow: ellipsis;
   `,
-
   ProductPrice: styled.span`
     font-size: 18px;
   `,
-
   CartButton: styled.button`
     background-color: transparent;
     border: none;
     cursor: pointer;
   `,
-
   CartSvg: styled.img`
     max-width: 100%;
   `,
