+++ src/components/ProductItem/index.js
@@ -1,0 +1,2 @@
+import ProductItem from './ProductItem';
+export default ProductItem;
