+++ src/components/Skeleton/index.js
@@ -1,0 +1,2 @@
+import Skeleton from './Skeleton';
+export default Skeleton;
