+++ src/pages/NotFound/index.js
@@ -1,0 +1,2 @@
+import NotFound from './NotFound';
+export default NotFound;
