+++ src/pages/Product/index.js
@@ -1,0 +1,2 @@
+import Product from './Product';
+export default Product;
