+++ src/pages/ProductList/index.js
@@ -1,0 +1,2 @@
+import ProductList from './ProductList';
+export default ProductList;
